components/train/pyproject.toml
[tool.poetry]
name = "train"
version = "0.1.0"
description = ""
authors = ["nomnomnonono <nozomu.koujiguchi1204@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
scikit-learn = "^1.3.0"
pandas = "^2.0.3"
joblib = "^1.3.2"
fsspec = "^2023.6.0"
gcsfs = "^2023.6.0"
python-dotenv = "^1.0.0"
lightgbm = "^4.1.0"
xgboost = "^2.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
